requirements.txt
pydantic==2.8.2
rpaframework==28.6.0
robocorp==1.4.0